requirements.txt
awscli==1.25.77
boto3==1.24.76
botocore==1.27.76
colorama==0.4.4
docutils==0.16
jmespath==1.0.1
pyasn1==0.4.8
python-dateutil==2.8.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
